+++ ui/components/DashboardComponent.js
@@ -25,8 +25,6 @@ import fetchControlPlanes from "./graphql/queries/ControlPlanesQuery";
 import fetchDataPlanes from "./graphql/queries/DataPlanesQuery";
 import fetchClusterResources from "./graphql/queries/ClusterResourcesQuery";
 import subscribeClusterResources from "./graphql/subscriptions/ClusterResourcesSubscription";
-import subscribeMeshModelSummary from "./graphql/subscriptions/MeshModelSummarySubscription";
-import fetchMeshModelSummary from "./graphql/queries/MeshModelSummaryQuery";
 import fetchAvailableNamespaces from "./graphql/queries/NamespaceQuery";
 import { submitPrometheusConfigure } from "./telemetry/prometheus/PrometheusComponent";
 import MUIDataTable from "mui-datatables";
@@ -132,9 +130,6 @@ class DashboardComponent extends React.Component {
       clusterResources : [],
       namespaceList : [],
       selectedNamespace : "default",
-      meshmodelSummarySelector : { type : "components" },
-      meshmodelSummarySelectorList : ["components", "relationships"],
-      meshmodelSummary : [],
 
       // subscriptions disposable
       dataPlaneSubscription : null,
@@ -142,8 +137,6 @@ class DashboardComponent extends React.Component {
       clusterResourcesSubscription : null,
       clusterResourcesQuery : null,
       namespaceQuery : null,
-      meshmodelSummarySubscription : null,
-      meshmodelSummaryQuery : null,
     };
   }
 
@@ -176,11 +169,6 @@ class DashboardComponent extends React.Component {
       this.state.controlPlaneSubscription.unsubscribe()
     }
     this.disposeWorkloadWidgetSubscription();
-  }
-
-  disposeMeshModelSummarySubscriptions = () => {
-    this.state.meshmodelSummarySubscription && this.state.meshmodelSummarySubscription.dispose();
-    this.state.meshmodelSummaryQuery && this.state.meshmodelSummaryQuery.unsubscribe();
   }
 
   initMeshSyncControlPlaneSubscription = () => {
@@ -266,38 +254,6 @@ class DashboardComponent extends React.Component {
     }
   }
 
-  initDashboardMeshModelSummaryQuery = () => {
-    const self = this;
-    let selector = self.state.meshmodelSummarySelector;
-
-    if (self._isMounted) {
-      // @ts-ignore
-      const meshmodelSummaryQuery = fetchMeshModelSummary(selector).subscribe({
-        next : (res) => {
-          this.setState({ meshmodelSummary : res?.meshmodelSummary })
-        },
-        error : (err) => console.log(err),
-      })
-
-      this.setState({ meshmodelSummaryQuery });
-    }
-  }
-
-  initDashboardMeshModelSummarySubscription = () => {
-    const self = this;
-    let selector = self.state.meshmodelSummarySelector;
-
-    if (self._isMounted) {
-      // @ts-ignore
-      const meshmodelSummarySubscription = subscribeMeshModelSummary((res) => {
-        this.setState({ meshmodelSummary : res?.meshmodelSummary })
-      }, {
-        selector : selector
-      });
-      this.setState({ meshmodelSummarySubscription });
-    }
-  }
-
   componentWillUnmount = () => {
     this._isMounted = false
     this.disposeSubscriptions()
@@ -316,8 +272,6 @@ class DashboardComponent extends React.Component {
       this.initDashboardClusterResourcesQuery();
       this.initDashboardClusterResourcesSubscription();
       this.initNamespaceQuery()
-      this.initDashboardMeshModelSummaryQuery();
-      this.initDashboardMeshModelSummarySubscription();
     }
   };
 
@@ -357,11 +311,6 @@ class DashboardComponent extends React.Component {
       this.initNamespaceQuery();
     }
 
-    if (prevState?.meshmodelSummarySelector !== this.state?.meshmodelSummarySelector) {
-      this.disposeMeshModelSummarySubscriptions();
-      this.initDashboardMeshModelSummaryQuery();
-      this.initDashboardMeshModelSummarySubscription();
-    }
   }
 
   getK8sClusterIds = () => {
@@ -670,10 +619,6 @@ class DashboardComponent extends React.Component {
       return `No resources detected in any of the cluster.`
     }
     return `No resources detected in the ${clusters.join(", ")} cluster(s).`
-  }
-
-  emptyStateMessageForMeshModelSummary = () => {
-    return "No MeshModel registered."
   }
 
   handleKubernetesClick = (id) => {
@@ -1115,128 +1060,6 @@ class DashboardComponent extends React.Component {
     return null;
   };
 
-   /**
-   * MeshModelSummaryCard takes in the meshmodel related data
-   * and renders a table with registered meshmodel information of
-   * the selected type of model (like relationships, components etc)
-   * @param {
-   * {
-   *   kind, count
-   * }[]
-   * } meshmodelSummary
-   */
-   MeshModelSummaryCard = (meshmodelSummary = []) => {
-     const self = this;
-     let kindSort = "asc";
-     let countSort = "asc";
-     const switchSortOrder = (type) => {
-       if (type === "kindSort") {
-         kindSort = (kindSort === "asc") ? "desc" : "asc";
-         countSort = "asc";
-       } else if (type === "countSort") {
-         countSort = (countSort === "asc") ? "desc" : "asc";
-         kindSort = "asc";
-       }
-     }
-
-     const columns = [
-       {
-         name : "name",
-         label : "Name",
-         options : {
-           filter : false,
-           sort : true,
-           searchable : true,
-           setCellProps : () => ({ style : { textAlign : "center" } }),
-           customHeadRender : ({ index, ...column }, sortColumn) => {
-             return (
-               <TableCell key={index} style={{ textAlign : "center" }} onClick={() => {
-                 sortColumn(index); switchSortOrder("kindSort");
-               }}>
-                 <TableSortLabel active={column.sortDirection != null} direction={kindSort} >
-                   <b>{column.label}</b>
-                 </TableSortLabel>
-               </TableCell>
-
-             )
-           }
-         },
-       },
-       {
-         name : "count",
-         label : "Count",
-         options : {
-           filter : false,
-           sort : true,
-           searchable : true,
-           setCellProps : () => ({ style : { textAlign : "center" } }),
-           customHeadRender : ({ index, ...column }, sortColumn) => {
-             return (
-               <TableCell key={index} style={{ textAlign : "center" }} onClick={() => {
-                 sortColumn(index); switchSortOrder("countSort");
-               }}>
-                 <TableSortLabel active={column.sortDirection != null} direction={countSort} >
-                   <b>{column.label}</b>
-                 </TableSortLabel>
-               </TableCell>
-
-             )
-           }
-         },
-       },
-     ]
-
-     const options = {
-       filter : false,
-       selectableRows : "none",
-       responsive : "scrollMaxHeight",
-       print : false,
-       download : false,
-       viewColumns : false,
-       pagination : false,
-       fixedHeader : true,
-       customToolbar : () => {
-         return (
-           <>
-             {self.state.meshmodelSummarySelectorList && (
-               <Select
-                 value={self.state.meshmodelSummarySelector.type}
-                 onChange={(e) =>
-                   self.setState({ meshmodelSummarySelector : { type : e.target.value } })
-                 }
-               >
-                 {self.state.meshmodelSummarySelectorList && self.state.meshmodelSummarySelectorList.map((opts) => <MenuItem key={opts} value={opts}>{opts}</MenuItem>)}
-               </Select>
-             )}
-           </>
-         )
-       }
-     }
-
-     if (Array.isArray(meshmodelSummary) && meshmodelSummary?.length)
-       return (
-         <Paper elevation={1} style={{ padding : "2rem" }}>
-           <MuiThemeProvider theme={this.getMuiTheme()}>
-             <MUIDataTable
-               title={
-                 <>
-                   <div style={{ display : "flex", alignItems : "center", marginBottom : "1rem" }}>
-                     <img src={"/static/img/all_mesh.svg"} className={this.props.classes.icon} style={{ marginRight : "0.75rem" }} />
-                     <Typography variant="h6">Registered MeshModel</Typography>
-                   </div>
-                 </>
-               }
-               data={meshmodelSummary}
-               options={options}
-               columns={columns}
-             />
-           </MuiThemeProvider>
-         </Paper>
-       );
-
-     return null;
-   };
-
   handlePrometheusClick = () => {
     this.props.updateProgress({ showProgress : true });
     const self = this;
@@ -1531,29 +1354,6 @@ class DashboardComponent extends React.Component {
       </>
     );
 
-    const showMeshModelSummary = (
-      <>
-        {self?.state?.meshmodelSummary[self?.state?.meshmodelSummarySelector?.type] && self?.state?.meshmodelSummary[self?.state?.meshmodelSummarySelector?.type].length > 0
-          ? (
-            self.MeshModelSummaryCard(self?.state?.meshmodelSummary[self?.state?.meshmodelSummarySelector?.type])
-          )
-          : (
-            <div
-              style={{
-                padding : "2rem",
-                display : "flex",
-                justifyContent : "center",
-                alignItems : "center",
-                flexDirection : "column",
-              }}
-            >
-              <Typography style={{ fontSize : "1.5rem", marginBottom : "2rem" }} align="center" color="textSecondary">
-                {this.emptyStateMessageForMeshModelSummary()}
-              </Typography>
-            </div>
-          )}
-      </>
-    );
     return (
       <NoSsr>
         <Popup />
@@ -1566,14 +1366,6 @@ class DashboardComponent extends React.Component {
                     Workloads
                   </Typography>
                   {showClusterResources}
-                </div>
-              </Grid>
-              <Grid item xs={12} md={12}>
-                <div className={classes.dashboardSection} data-test="workloads">
-                  <Typography variant="h6" gutterBottom className={classes.chartTitle}>
-                    MeshModel
-                  </Typography>
-                  {showMeshModelSummary}
                 </div>
               </Grid>
               <Grid item xs={12} md={12}>
