+++ ui/components/MesherySettings.js
@@ -6,9 +6,8 @@ import { connect } from 'react-redux';
 import { bindActionCreators } from "redux";
 import Tabs from '@material-ui/core/Tabs';
 import Tab from '@material-ui/core/Tab';
-import Typography from '@material-ui/core/Typography';
 import {
-  AppBar, Paper, Tooltip, IconButton, Button
+  AppBar, Paper, Tooltip, Button, IconButton, MenuItem, Select, TableCell, TableSortLabel, Typography
 } from '@material-ui/core';
 import CloseIcon from "@material-ui/icons/Close";
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
@@ -28,6 +27,11 @@ import { ctxUrl } from '../utils/multi-ctx';
 import PromptComponent from './PromptComponent';
 import resetDatabase from './graphql/queries/ResetDatabaseQuery';
 import { iconMedium } from '../css/icons.styles';
+import subscribeMeshModelSummary from "./graphql/subscriptions/MeshModelSummarySubscription";
+import fetchMeshModelSummary from "./graphql/queries/MeshModelSummaryQuery";
+import { MuiThemeProvider, createTheme } from '@material-ui/core/styles';
+import MUIDataTable from "mui-datatables";
+
 
 const styles = (theme) => ({
   wrapperClss : {
@@ -77,7 +81,19 @@ const styles = (theme) => ({
     justifyContent : "space-between",
     paddingLeft : "1rem",
     maxWidth : "90%"
-  }
+  },
+  dashboardSection : {
+    backgroundColor : "#fff",
+    padding : theme.spacing(2),
+    borderRadius : 4,
+    height : "100%",
+  },
+  cardHeader : { fontSize : theme.spacing(2), },
+  card : {
+    height : "100%",
+    marginTop : theme.spacing(2),
+  },
+  cardContent : { height : "100%", },
 });
 
 function TabContainer(props) {
@@ -100,6 +116,7 @@ class MesherySettings extends React.Component {
       k8sconfig, meshAdapters, grafana, prometheus, router : { asPath }
     } = props;
 
+    this._isMounted = false;
     let tabVal = 0, subTabVal = 0;
     const splittedPath = asPath.split('#');
     if (splittedPath.length >= 2 && splittedPath[1]) {
@@ -151,7 +168,13 @@ class MesherySettings extends React.Component {
       // Array of scanned prometheus urls
       scannedPrometheus : [],
       // Array of scanned grafan urls
-      scannedGrafana : []
+      scannedGrafana : [],
+
+      meshmodelSummarySelector : { type : "components" },
+      meshmodelSummarySelectorList : ["components", "relationships"],
+      meshmodelSummary : [],
+      meshmodelSummarySubscription : null,
+      meshmodelSummaryQuery : null,
     };
 
     this.systemResetRef = React.createRef();
@@ -170,13 +193,68 @@ class MesherySettings extends React.Component {
     return null;
   }
 
-  componentDidMount() {
-    this.fetchPromGrafanaScanData();
+  disposeMeshModelSummarySubscriptions = () => {
+    this.state.meshmodelSummarySubscription && this.state.meshmodelSummarySubscription.dispose();
+    this.state.meshmodelSummaryQuery && this.state.meshmodelSummaryQuery.unsubscribe();
   }
 
-  componentDidUpdate(prevProps) {
+  initDashboardMeshModelSummaryQuery = () => {
+    const self = this;
+    let selector = self.state.meshmodelSummarySelector;
+
+    if (self._isMounted) {
+      // @ts-ignore
+      const meshmodelSummaryQuery = fetchMeshModelSummary(selector).subscribe({
+        next : (res) => {
+          this.setState({ meshmodelSummary : res?.meshmodelSummary })
+        },
+        error : (err) => console.log(err),
+      })
+
+      this.setState({ meshmodelSummaryQuery });
+    }
+  }
+
+  initDashboardMeshModelSummarySubscription = () => {
+    const self = this;
+    let selector = self.state.meshmodelSummarySelector;
+
+    if (self._isMounted) {
+      // @ts-ignore
+      const meshmodelSummarySubscription = subscribeMeshModelSummary((res) => {
+        this.setState({ meshmodelSummary : res?.meshmodelSummary })
+      }, {
+        selector : selector
+      });
+      this.setState({ meshmodelSummarySubscription });
+    }
+  }
+
+  disposeSubscriptions = () => {
+    this.disposeMeshModelSummarySubscriptions()
+  }
+
+  componentDidMount() {
+    this._isMounted = true
+
+    this.fetchPromGrafanaScanData();
+    this.disposeSubscriptions()
+
+    if (this._isMounted) {
+      this.initDashboardMeshModelSummaryQuery();
+      this.initDashboardMeshModelSummarySubscription();
+    }
+  }
+
+  componentDidUpdate(prevProps, prevState) {
     if (prevProps.selectedK8sContexts.length != this.props.selectedK8sContexts.length) {
       this.fetchPromGrafanaScanData();
+    }
+
+    if (prevState?.meshmodelSummarySelector !== this.state?.meshmodelSummarySelector) {
+      this.disposeMeshModelSummarySubscriptions();
+      this.initDashboardMeshModelSummaryQuery();
+      this.initDashboardMeshModelSummarySubscription();
     }
   }
 
@@ -206,6 +284,206 @@ class MesherySettings extends React.Component {
       self.handleError("Unable to fetch Prometheus and Grafana details")
     )
   }
+
+  emptyStateMessageForMeshModelSummary = () => {
+    return "No MeshModel registered."
+  }
+
+  getMuiTheme = () => createTheme({
+    shadows : ["none"],
+    overrides : {
+      MUIDataTable : {
+      },
+      MuiInput : {
+        underline : {
+          "&:hover:not(.Mui-disabled):before" : {
+            borderBottom : "2px solid #222"
+          },
+          "&:after" : {
+            borderBottom : "2px solid #222"
+          }
+        }
+      },
+      MUIDataTableSearch : {
+        searchIcon : {
+          color : "#607d8b",
+          marginTop : "7px",
+          marginRight : "8px",
+        },
+        clearIcon : {
+          "&:hover" : {
+            color : "#607d8b"
+          }
+        },
+      },
+      MUIDataTableSelectCell : {
+        checkboxRoot : {
+          '&$checked' : {
+            color : '#607d8b',
+          },
+        },
+      },
+      MUIDataTableToolbar : {
+        iconActive : {
+          color : "#222"
+        },
+        icon : {
+          "&:hover" : {
+            color : "#607d8b"
+          }
+        },
+      },
+    }
+  })
+
+     /**
+   * MeshModelSummaryCard takes in the meshmodel related data
+   * and renders a table with registered meshmodel information of
+   * the selected type of model (like relationships, components etc)
+   * @param {
+   * {
+   *   kind, count
+   * }[]
+   * } meshmodelSummary
+   */
+     MeshModelSummaryCard = (meshmodelSummary = []) => {
+       const self = this;
+       let kindSort = "asc";
+       let countSort = "asc";
+       const switchSortOrder = (type) => {
+         if (type === "kindSort") {
+           kindSort = (kindSort === "asc") ? "desc" : "asc";
+           countSort = "asc";
+         } else if (type === "countSort") {
+           countSort = (countSort === "asc") ? "desc" : "asc";
+           kindSort = "asc";
+         }
+       }
+
+       const columns = [
+         {
+           name : "name",
+           label : "Name",
+           options : {
+             filter : false,
+             sort : true,
+             searchable : true,
+             setCellProps : () => ({ style : { textAlign : "center" } }),
+             customHeadRender : ({ index, ...column }, sortColumn) => {
+               return (
+                 <TableCell key={index} style={{ textAlign : "center" }} onClick={() => {
+                   sortColumn(index); switchSortOrder("kindSort");
+                 }}>
+                   <TableSortLabel active={column.sortDirection != null} direction={kindSort} >
+                     <b>{column.label}</b>
+                   </TableSortLabel>
+                 </TableCell>
+
+               )
+             }
+           },
+         },
+         {
+           name : "count",
+           label : "Count",
+           options : {
+             filter : false,
+             sort : true,
+             searchable : true,
+             setCellProps : () => ({ style : { textAlign : "center" } }),
+             customHeadRender : ({ index, ...column }, sortColumn) => {
+               return (
+                 <TableCell key={index} style={{ textAlign : "center" }} onClick={() => {
+                   sortColumn(index); switchSortOrder("countSort");
+                 }}>
+                   <TableSortLabel active={column.sortDirection != null} direction={countSort} >
+                     <b>{column.label}</b>
+                   </TableSortLabel>
+                 </TableCell>
+
+               )
+             }
+           },
+         },
+       ]
+
+       const options = {
+         filter : false,
+         selectableRows : "none",
+         responsive : "scrollMaxHeight",
+         print : false,
+         download : false,
+         viewColumns : false,
+         pagination : false,
+         fixedHeader : true,
+         customToolbar : () => {
+           return (
+             <>
+               {self.state.meshmodelSummarySelectorList && (
+                 <Select
+                   value={self.state.meshmodelSummarySelector.type}
+                   onChange={(e) =>
+                     self.setState({ meshmodelSummarySelector : { type : e.target.value } })
+                   }
+                 >
+                   {self.state.meshmodelSummarySelectorList && self.state.meshmodelSummarySelectorList.map((opts) => <MenuItem key={opts} value={opts}>{opts}</MenuItem>)}
+                 </Select>
+               )}
+             </>
+           )
+         }
+       }
+
+       if (Array.isArray(meshmodelSummary) && meshmodelSummary?.length)
+         return (
+           <Paper elevation={1} style={{ padding : "2rem" }}>
+             <MuiThemeProvider theme={this.getMuiTheme()}>
+               <MUIDataTable
+                 title={
+                   <>
+                     <div style={{ display : "flex", alignItems : "center", marginBottom : "1rem" }}>
+                       <img src={"/static/img/all_mesh.svg"} className={this.props.classes.icon} style={{ marginRight : "0.75rem" }} />
+                       <Typography variant="h6">Registered MeshModel</Typography>
+                     </div>
+                   </>
+                 }
+                 data={meshmodelSummary}
+                 options={options}
+                 columns={columns}
+               />
+             </MuiThemeProvider>
+           </Paper>
+         );
+
+       return null;
+     };
+
+    showMeshModelSummary = () => {
+      const self = this;
+      return (
+        <>
+          {self?.state?.meshmodelSummary[self?.state?.meshmodelSummarySelector?.type] && self?.state?.meshmodelSummary[self?.state?.meshmodelSummarySelector?.type].length > 0
+            ? (
+              self.MeshModelSummaryCard(self?.state?.meshmodelSummary[self?.state?.meshmodelSummarySelector?.type])
+            )
+            : (
+              <div
+                style={{
+                  padding : "2rem",
+                  display : "flex",
+                  justifyContent : "center",
+                  alignItems : "center",
+                  flexDirection : "column",
+                }}
+              >
+                <Typography style={{ fontSize : "1.5rem", marginBottom : "2rem" }} align="center" color="textSecondary">
+                  {this.emptyStateMessageForMeshModelSummary()}
+                </Typography>
+              </div>
+            )}
+        </>
+      );
+    };
 
   /**
    * extractURLFromScanData scans the ingress urls from the
@@ -284,6 +562,9 @@ class MesherySettings extends React.Component {
             break;
           case 3:
             newRoute += '#system'
+            break;
+          case 4:
+            newRoute += '#meshmodel-summary'
           // case 3:
           //   newRoute += '#performance'
           //   break;
@@ -423,6 +704,16 @@ class MesherySettings extends React.Component {
                 tab="systemReset"
               />
             </Tooltip>
+            <Tooltip title="MeshModel Summary" placement="top">
+              <Tab
+                className={classes.tab}
+                icon={
+                  <FontAwesomeIcon icon={faDatabase}  style={iconMedium} />
+                }
+                label="MeshModel Summary"
+                tab="meshmodelSummary"
+              />
+            </Tooltip>
 
             {/*NOTE: Functionality of performance tab will be modified, until then keeping it and the related code commented */}
 
@@ -501,6 +792,18 @@ class MesherySettings extends React.Component {
               >
                 <Typography> System Reset </Typography>
               </Button>
+            </div>
+          </TabContainer>
+        )}
+        {tabVal === 4 && (
+          <TabContainer>
+            <div className={classes.container}>
+              <div className={classes.dashboardSection} data-test="workloads">
+                <Typography variant="h6" gutterBottom className={classes.chartTitle}>
+                    MeshModel
+                </Typography>
+                {this.showMeshModelSummary()}
+              </div>
             </div>
           </TabContainer>
         )}
